Break LoaderHeap codebase into multiple files
