+++ src/coreclr/utilcode/CMakeLists.txt
@@ -28,6 +28,10 @@ set(UTILCODE_COMMON_SOURCES
   guidfromname.cpp
   memorypool.cpp
   loaderheap.cpp
+  loaderheap_shared.cpp
+  explicitcontrolloaderheap.cpp
+  allocmemtracker.cpp
+  rangelist.cpp
   outstring.cpp
   ilformatter.cpp
   opinfo.cpp
